Fix stale references, unused imports, and naming inconsistencies
- Fix inference.py: dtype -> torch_dtype (prevents fp32 fallback)
- Remove unused imports: warnings, tempfile, SUBDIRS, CLASS_TO_CODE
- Remove dead code: label_to_code, duplicated _imread_unicode
- Rename firetrace -> flashback in all output filenames and dataset names
- Fix stale file references: visualize_origin.py, fire_detection.py, S5
- Add missing deps to requirements.txt: streamlit, plotly, Pillow
- Fix README: venv activation, scene counts (FLAME 6/SMOKE 2/NORMAL 3)
- Rename reports/ output files from firetrace_* to flashback_*

+++ src/core/inference.py
@@ -7,7 +7,6 @@ Uses PyAV backend for video decoding (Windows-compatible).
 import json
 import re
 import time
-import warnings
 import torch
 import transformers
 from pathlib import Path
@@ -54,7 +53,7 @@ class Reason2Model:
         self.model_name = model_name
         self.model = transformers.Qwen3VLForConditionalGeneration.from_pretrained(
             model_name,
-            dtype=dtype,
+            torch_dtype=dtype,
             device_map=device_map,
             attn_implementation="sdpa",
         )
